refactor: create useDepositMax hook
